Add Solce API notifiers to handle API deletion/lifecycle change
