+++ components/apimgt/org.wso2.carbon.apimgt.rest.api.admin.v1/src/gen/java/org/wso2/carbon/apimgt/rest/api/admin/v1/dto/AdditionalPropertyDTO.java
@@ -1,4 +1,4 @@
-package org.wso2.carbon.apimgt.rest.api.publisher.v1.dto;
+package org.wso2.carbon.apimgt.rest.api.admin.v1.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonCreator;
@@ -19,82 +19,83 @@ import javax.validation.Valid;
 
 
 public class AdditionalPropertyDTO   {
-
+  
     private String key = null;
     private String value = null;
 
-    /**
-     **/
-    public AdditionalPropertyDTO key(String key) {
-        this.key = key;
-        return this;
+  /**
+   **/
+  public AdditionalPropertyDTO key(String key) {
+    this.key = key;
+    return this;
+  }
+
+  
+  @ApiModelProperty(example = "Organization", value = "")
+  @JsonProperty("key")
+  public String getKey() {
+    return key;
+  }
+  public void setKey(String key) {
+    this.key = key;
+  }
+
+  /**
+   **/
+  public AdditionalPropertyDTO value(String value) {
+    this.value = value;
+    return this;
+  }
+
+  
+  @ApiModelProperty(example = "wso2", value = "")
+  @JsonProperty("value")
+  public String getValue() {
+    return value;
+  }
+  public void setValue(String value) {
+    this.value = value;
+  }
+
+
+  @Override
+  public boolean equals(java.lang.Object o) {
+    if (this == o) {
+      return true;
     }
-
-
-    @ApiModelProperty(example = "Organization", value = "")
-    @JsonProperty("key")
-    public String getKey() {
-        return key;
+    if (o == null || getClass() != o.getClass()) {
+      return false;
     }
-    public void setKey(String key) {
-        this.key = key;
+    AdditionalPropertyDTO additionalProperty = (AdditionalPropertyDTO) o;
+    return Objects.equals(key, additionalProperty.key) &&
+        Objects.equals(value, additionalProperty.value);
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hash(key, value);
+  }
+
+  @Override
+  public String toString() {
+    StringBuilder sb = new StringBuilder();
+    sb.append("class AdditionalPropertyDTO {\n");
+    
+    sb.append("    key: ").append(toIndentedString(key)).append("\n");
+    sb.append("    value: ").append(toIndentedString(value)).append("\n");
+    sb.append("}");
+    return sb.toString();
+  }
+
+  /**
+   * Convert the given object to string with each line indented by 4 spaces
+   * (except the first line).
+   */
+  private String toIndentedString(java.lang.Object o) {
+    if (o == null) {
+      return "null";
     }
-
-    /**
-     **/
-    public AdditionalPropertyDTO value(String value) {
-        this.value = value;
-        return this;
-    }
-
-
-    @ApiModelProperty(example = "wso2", value = "")
-    @JsonProperty("value")
-    public String getValue() {
-        return value;
-    }
-    public void setValue(String value) {
-        this.value = value;
-    }
-
-
-    @Override
-    public boolean equals(java.lang.Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        AdditionalPropertyDTO additionalProperty = (AdditionalPropertyDTO) o;
-        return Objects.equals(key, additionalProperty.key) &&
-                Objects.equals(value, additionalProperty.value);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(key, value);
-    }
-
-    @Override
-    public String toString() {
-        StringBuilder sb = new StringBuilder();
-        sb.append("class AdditionalPropertyDTO {\n");
-
-        sb.append("    key: ").append(toIndentedString(key)).append("\n");
-        sb.append("    value: ").append(toIndentedString(value)).append("\n");
-        sb.append("}");
-        return sb.toString();
-    }
-
-    /**
-     * Convert the given object to string with each line indented by 4 spaces
-     * (except the first line).
-     */
-    private String toIndentedString(java.lang.Object o) {
-        if (o == null) {
-            return "null";
-        }
-        return o.toString().replace("\n", "\n    ");
-    }
+    return o.toString().replace("\n", "\n    ");
+  }
 }
+
