+++ components/apimgt/org.wso2.carbon.apimgt.rest.api.publisher.v1/src/main/java/org/wso2/carbon/apimgt/rest/api/publisher/v1/impl/ApisApiServiceImpl.java
@@ -4445,20 +4445,13 @@ public class ApisApiServiceImpl implements ApisApiService {
         String organization = RestApiUtil.getValidatedOrganization(messageContext);
 
         Map<String, Environment> environments = APIUtil.getEnvironments();
-        Map<String, Environment> gatewayEnvironments = APIUtil.getReadOnlyGatewayEnvironments();
         List<APIRevisionDeployment> apiRevisionDeployments = new ArrayList<>();
         for (APIRevisionDeploymentDTO apiRevisionDeploymentDTO : apIRevisionDeploymentDTOList) {
             APIRevisionDeployment apiRevisionDeployment = new APIRevisionDeployment();
             apiRevisionDeployment.setRevisionUUID(revisionId);
             String environment = apiRevisionDeploymentDTO.getName();
             if (environments.get(environment) == null) {
-                if (gatewayEnvironments != null) {
-                    if (gatewayEnvironments.get(environment) == null) {
-                        RestApiUtil.handleBadRequest("Gateway environment not found: " + environment, log);
-                    }
-                } else {
-                    RestApiUtil.handleBadRequest("Gateway environment not found: " + environment, log);
-                }
+                RestApiUtil.handleBadRequest("Gateway environment not found: " + environment, log);
             }
             apiRevisionDeployment.setDeployment(environment);
             apiRevisionDeployment.setVhost(apiRevisionDeploymentDTO.getVhost());
@@ -4522,7 +4515,6 @@ public class ApisApiServiceImpl implements ApisApiService {
         }
 
         Map<String, Environment> environments = APIUtil.getEnvironments();
-        Map<String, Environment> gatewayEnvironments = APIUtil.getReadOnlyGatewayEnvironments();
         List<APIRevisionDeployment> apiRevisionDeployments = new ArrayList<>();
         if (allEnvironments) {
             apiRevisionDeployments = apiProvider.getAPIRevisionDeploymentList(revisionId);
@@ -4532,13 +4524,7 @@ public class ApisApiServiceImpl implements ApisApiService {
                 apiRevisionDeployment.setRevisionUUID(revisionId);
                 String environment = apiRevisionDeploymentDTO.getName();
                 if (environments.get(environment) == null) {
-                    if (gatewayEnvironments != null) {
-                        if (gatewayEnvironments.get(environment) == null) {
-                            RestApiUtil.handleBadRequest("Gateway environment not found: " + environment, log);
-                        }
-                    } else {
-                        RestApiUtil.handleBadRequest("Gateway environment not found: " + environment, log);
-                    }
+                    RestApiUtil.handleBadRequest("Gateway environment not found: " + environment, log);
                 }
                 apiRevisionDeployment.setDeployment(environment);
                 apiRevisionDeployment.setVhost(apiRevisionDeploymentDTO.getVhost());
